refactor: 총동연 공지 공통 컴포넌트 분리하여 구조 개선 및 모드에 따라 분기 (#30)
