+++ src/pages/admin/union/notice/AdminUnionNoticeEditPage.tsx
@@ -1,0 +1,5 @@
+import { UnionNoticeRegisterForm } from '@/components/UnionNotice';
+
+export default function AdminUnionNoticeEditPage() {
+    return <UnionNoticeRegisterForm mode="edit" />;
+}
