+++ src/pages/admin/union/notice/AdminUnionNoticePage.tsx
@@ -1,11 +1,8 @@
-import { UnionNotice } from '@/components/UnionNotice/UnionNotice';
-import { useLocation } from 'react-router-dom';
+import { UnionNoticeList } from '@/components/UnionNotice/UnionNoticeList';
 
 const AdminUnionNoticePage = () => {
-    const location = useLocation();
-    const isEditMode = location.pathname.includes('/admin/union/notice');
-
-    return <UnionNotice editMode={isEditMode} />;
+    // 등록 모드 (어드민 총동연에서의 공지사항 등록 가능)
+    return <UnionNoticeList mode="register" />;
 };
 
 export { AdminUnionNoticePage };
