+++ src/pages/notice/UnionNoticePage.tsx
@@ -1,11 +1,8 @@
-import { UnionNotice } from '@/components/UnionNotice/UnionNotice';
-import { useLocation } from 'react-router-dom';
+import { UnionNoticeList } from '@/components/UnionNotice/UnionNoticeList';
 
 const UnionNoticePage = () => {
-    const location = useLocation();
-    const isEditMode = location.pathname.includes('/admin/union/notice');
-
-    return <UnionNotice editMode={isEditMode} />;
+    // 읽기 모드 (어드민 총동연에서의 공지사항 등록 불가)
+    return <UnionNoticeList mode="read" />;
 };
 
 export { UnionNoticePage };
